backend/python-services/incident_integrator/requirements.txt
fastapi
uvicorn
python-dotenv
pytest
httpx
pydantic
requests
jira
PyGithub 
mypy 